requirements.txt
groq
python-dotenv
langchain-groq
flask
twilio
crewai
fastapi
uvicorn